+++ luag-lib/v2/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## version 2.2
 1. `time` and `date` functions
+2. `write` now has a 'scale' optional parameter
 
 ## version 2.1
 1. key functions now accept key names as arguments